success

+++ app.py
@@ -1,11 +1,13 @@
 # Using flask to make an api
 # import necessary libraries and functions
-from bson.objectid import ObjectId
-from flask import Flask, jsonify, request
+from flask import Flask, request
+from bson.json_util import dumps
 from flask_pymongo import PyMongo
+from bson import json_util, ObjectId
+
 app = Flask(__name__)
 # connecting with DB
-uname,pwd,dbName = "keshav","forgotpassword","lms"
+uname,pwd,dbName = "keshav","forgotpassword","Library"
 uri = "mongodb+srv://"+uname+":"+pwd+"@keshav-lms.biuic.mongodb.net/"+dbName+"?retryWrites=true&w=majority"
 mongodb_client = PyMongo(app, uri)
 db = mongodb_client.db
@@ -17,7 +19,7 @@ def home():
 		return ("Welcome to Keshav's Library Management System. Use CRUD operations with REST Api" ,202)
 
 
-@app.route('/LMS/addBooks', methods = ['POST'])
+@app.route('/LMS/addBook', methods = ['POST'])
 def addBook():
 	body = request.json
 	author, bookName = body["author"], body["bookName"]
@@ -43,23 +45,23 @@ def getBooks():
 	id, author, bookName = args.get("id"), args.get("author"),args.get("bookName")
 	print(id,author,bookName)
 	if(id):
-		book = db.books.find_one({'_id': ObjectId(id)})
-		return (jsonify(book), 200)
+		return (dumps(db.books.find_one({'_id': ObjectId(id)}),default=json_util.default), 200)
 	elif(author):
-		book = db.books.find_one({'author': author})
-		return (jsonify(book), 200)
+		return (dumps(db.books.find_one({'author': author}), default=json_util.default),200)
 	elif(bookName):
-		book = db.books.find_one({'bookName': bookName})
-		return (jsonify(book), 200)
+		return (dumps(db.books.find_one({'bookName': bookName}), default=json_util.default),200)
+
 	else:
-		books = db.books.find()
-		return (jsonify(books), 200)
+		return (dumps(db.books.find(), default=json_util.default),200)
 
 @app.route("/LMS/deleteBook",methods=['DELETE'])
 def deleteBook():
 	id = request.json['id']
-	db.book.delete_one({'_id':ObjectId(id)})
-	return ("Book deleted!!",200)
+	if( db.books.find_one({'_id': ObjectId(id)}) ):
+		db.book.delete_one({'_id': ObjectId(id)})
+		return ("Book deleted!!",200)
+	else:
+		return ("Invalid ID",404)
 
 @app.errorhandler(404)
 def not_found(error=None):
@@ -70,4 +72,4 @@ def not_found(error=None):
 
 # driver function
 if __name__ == '__main__':
-	app.run(debug = True,port=80)
+	app.run(host="0.0.0.0",debug = True,port=80)
